backend/requirements.txt
flask
flask-cors
google-generativeai
pandas
sqlalchemy
pyodbc
python-dotenv
requests
